+++ AndroidApp/app/src/main/java/com/zerosepaisa/liferesetos/navigation/bottomnav/BottomNavigationBar.kt
@@ -5,35 +5,43 @@ import androidx.compose.material3.NavigationBar
 import androidx.compose.material3.NavigationBarItem
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.navigation.NavDestination.Companion.hierarchy
+import androidx.navigation.NavHostController
+import androidx.navigation.compose.currentBackStackEntryAsState
 
 @Composable
 fun BottomNavigationBar(
-
-    selectedRoute: String,
-
-    onItemClick: (String) -> Unit
-
+    navController: NavHostController
 ) {
 
     val items = listOf(
-
         BottomNavItem.Home,
         BottomNavItem.Journey,
         BottomNavItem.Focus,
         BottomNavItem.Profile
-
     )
+
+    val navBackStackEntry by navController.currentBackStackEntryAsState()
+    val currentDestination = navBackStackEntry?.destination
 
     NavigationBar {
 
         items.forEach { item ->
 
+            val selected = currentDestination?.hierarchy?.any { it.route == item.route } == true
+
             NavigationBarItem(
 
-                selected = selectedRoute == item.route,
+                selected = selected,
 
                 onClick = {
-                    onItemClick(item.route)
+                    if (!selected) {
+                        navController.navigate(item.route) {
+                            popUpTo(0) { inclusive = true }
+                            launchSingleTop = true
+                        }
+                    }
                 },
 
                 icon = {